build: fix linking with libtiff with lerc support
Lerc is written in C++.  When C and C++ objects are linked, a C++
linker should be used to ensure C++-specific things are correctly
handled.  See e.g. this comment in the Meson source for reference[1].
One symptom of using a C linker to link with C++ objects is that
libstdc++ won't be linked when building static executables, causing
link failures.

Unfortunately, Meson does not know whether dependencies found by
pkg-config are C++, and therefore require a C++ linker, so we have to
tell it ourselves to use a C++ linker.  There's no way to check
whether libtiff is built with Lerc support, so we always use a C++
linker if one is available and libtiff support is enabled.  If a C++
linker ends up being used to link only C objects, it shouldn't do any
harm.

[1]: https://github.com/mesonbuild/meson/blob/9fd5281befe7881c9d1210c9e6865382bc0f2b08/mesonbuild/build.py#L1558-L1565